requirements.txt
ollama
langchain>=0.2.0
langchain-core>=0.2.0
whisper-openai
pillow
pyaudio